requirements.txt
langchain-google-genai>=0.1.4
browser-use>=0.1.9
python-dotenv>=1.0.1
pydantic>=2.6.3
